backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
chromadb
pandas
openpyxl
langchain
langchain-community
tiktoken